Detach wgl and win32
